se arregla ventas y historial de ventas

+++ delusuario/src/main/java/com/hora/delusuario/repository/HistorialVentaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HistorialVentaRepository extends JpaRepository<HistorialVentaEntity, Long> {
+public interface HistorialVentaRepository extends JpaRepository<HistorialVentaEntity, Integer> {
     // Aquí puedes agregar métodos de consulta personalizados si los necesitas
 }
